Maria branch

+++ backend/queries/owners.js
@@ -1,36 +1,42 @@
 const db = require("../db/index");
 
 const signUp = async (req, res, next) => {
-  let {user_id, email} = req.body;
+  let { user_id, email } = req.body;
   try {
-    let user = await db.one("INSERT INTO owners (user_id, email) VALUES($1, $2) RETURNING *", [user_id, email]);
+    let user = await db.one(
+      "INSERT INTO owners (user_id, email) VALUES($1, $2) RETURNING *",
+      [user_id, email]
+    );
     res.status(200).json({
       status: "success",
       message: "created user",
-      payload: user
+      payload: user,
     });
   } catch (err) {
     res.status(400).json({
       status: "Error",
       message: "could not create new user",
-      payload: err
+      payload: err,
     });
     next();
   }
 };
 const getSingleOwner = async (req, res, next) => {
   try {
-    let owner = await db.one(`SELECT * FROM owners WHERE user_id =${req.params.id}`);
+    let owner = await db.one(
+      `SELECT * FROM owners WHERE user_id =$1`,
+      req.params.id
+    );
     res.status(200).json({
       status: "success",
       message: "single owner",
-      payload: owner
+      payload: owner,
     });
   } catch (err) {
     res.status(400).json({
       status: "Error",
       message: "Error get single Owner",
-      payload: err
+      payload: err,
     });
     next();
   }
@@ -38,38 +44,38 @@ const getSingleOwner = async (req, res, next) => {
 const deleteOwner = async (req, res, next) => {
   try {
     let { id } = req.params;
-    let owner = await db.none(
-      "DELETE FROM owners WHERE id =$1",
-      id
-    );
+    let owner = await db.none("DELETE FROM owners WHERE id =$1", id);
     res.status(200).json({
       status: "success",
       message: "deleted owner",
-      payload: owner
+      payload: owner,
     });
   } catch (err) {
     res.status(400).json({
       status: "Error",
       message: "Error delete owner",
-      payload: err
+      payload: err,
     });
     next();
   }
 };
 const createOwner = async (req, res, next) => {
   try {
-    let { owner_id, owner_name } = req.body
-    let owner = await db.one("INSERT INTO owners (owner_id, owner_name) VALUES ($1,$2) RETURNING *", [owner_id, owner_name]);
+    let { owner_id, owner_name } = req.body;
+    let owner = await db.one(
+      "INSERT INTO owners (owner_id, owner_name) VALUES ($1,$2) RETURNING *",
+      [owner_id, owner_name]
+    );
     res.status(200).json({
       status: "success",
       message: "created owner",
-      payload: owner
+      payload: owner,
     });
   } catch (err) {
     res.status(400).json({
       status: "Error",
       message: "Error for create Owner",
-      payload: err
+      payload: err,
     });
     next();
   }
@@ -85,15 +91,21 @@ const editOwner = async (req, res, next) => {
     res.status(200).json({
       status: "success",
       message: "updated owner",
-      payload: owner
+      payload: owner,
     });
   } catch (err) {
     res.status(400).json({
       status: "Error",
       message: "Error edit Owner",
-      payload: err
+      payload: err,
     });
     next();
   }
 };
-module.exports = { signUp, getSingleOwner, createOwner, editOwner, deleteOwner };
+module.exports = {
+  signUp,
+  getSingleOwner,
+  createOwner,
+  editOwner,
+  deleteOwner,
+};
